requirements.txt
requests
beautifulsoup4
lxml
gspread
google-auth
